night mode on

+++ backend/controllers/product.contoller.js
@@ -50,11 +50,17 @@ export const getProducts =  async (req, res) => {
 
   export const removeProduct = async (req, res) => {
     const { id } = req.params;
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+      return res.status(404).json({
+        success: false,
+        message: "Product not found",
+      });
+    }
     try {
       await Product.findByIdAndDelete(id);
       res.status(200).json({ success: true, message: "product deleted" });
     } catch (error) {
       console.error("error deleting item");
-      res.status(404).json({ success: false, message: "couldn't delete item" });
+      res.status(500).json({ success: false, message: "server error" });
     }
   }
